+++ 1.Distributed-Systems/app/api/controller/messages/health.py
@@ -1,6 +1,6 @@
 import time
 
-from healthy import HealthStatus
+from service.healthy import HealthStatus
 
 # TODO: actually it it should be smartest way - check env. and so on, but it is out of that scope
 def get_health():
