+++ 1.Distributed-Systems/app/api/data/messages/__init__.py
@@ -1,0 +1,2 @@
+from api.data.messages.create import CreateMessageRequestData
+
