--- 1.Distributed-Systems/app/request_dto/messages/__init__.py
@@ -1,2 +1,0 @@
-from request_dto.messages.create import CreateMessageRequestData
-
